Create Hi, I'm @lancebatchelor
I'm a Biologist that is interested and educated in many disciplines of science. I specialize in molecular biology. 
I am for critical thinking and skepticism.
I am an atheist and I am interested in how religious people do not use critical thinking and evidence in their belief system. Faith is not a pathway to truth.